feat: added error handling
